backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
langchain-core==0.3.72
langchain==0.3.27
ollama
langchain-openai==0.2.12
openai>=1.0.0
lark-oapi==1.5.2
kubernetes==31.0.0
requests==2.32.3
jinja2==3.1.5
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0
aiosqlite>=0.20
sqlmodel>=0.0.16
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
